pyproject.toml
[project]
name = "multipage-dashboard-streamlit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas-stubs~=2.3.3",
    "plotly>=6.5.2",
    "streamlit>=1.53.0",
]
